requirements.txt
telethon==1.35.0
python-dotenv==1.0.1
aiogram==3.10.0
fastapi==0.112.0
uvicorn==0.30.0
aiohttp==3.10.0
pydantic==2.8.0
async-lru==2.0.4
